[Chore]Move the Currency Formatter to the viewmodel
